+++ js/protocol/Recording.d.ts
@@ -4,14 +4,6 @@ import { ScreenShot } from "./Graphics";
  * Globally unique identifier for a recording.
  */
 export declare type RecordingId = string;
-export interface RecordingData {
-    id: RecordingId;
-    url: string;
-    title: string;
-    duration: number;
-    lastScreenData: string;
-    lastScreenMimeType: string;
-}
 /**
  * Unique identifier for the software used to produce a recording.
  */
