+++ json/protocol.json
@@ -14,36 +14,6 @@
                     "type": "string"
                 },
                 {
-                    "id": "RecordingData",
-                    "type": "object",
-                    "properties": [
-                        {
-                            "name": "id",
-                            "$ref": "RecordingId"
-                        },
-                        {
-                            "name": "url",
-                            "type": "string"
-                        },
-                        {
-                            "name": "title",
-                            "type": "string"
-                        },
-                        {
-                            "name": "duration",
-                            "type": "integer"
-                        },
-                        {
-                            "name": "lastScreenData",
-                            "type": "string"
-                        },
-                        {
-                            "name": "lastScreenMimeType",
-                            "type": "string"
-                        }
-                    ]
-                },
-                {
                     "id": "BuildId",
                     "description": "Unique identifier for the software used to produce a recording.",
                     "type": "string"
@@ -3028,7 +2998,7 @@
                     "parameters": [
                         {
                             "name": "recordingData",
-                            "$ref": "Recording.RecordingData"
+                            "$ref": "RecordingData"
                         },
                         {
                             "name": "authId",
@@ -3224,6 +3194,36 @@
                         {
                             "name": "checksum",
                             "description": "Checksum of the resource's contents.",
+                            "type": "string"
+                        }
+                    ]
+                },
+                {
+                    "id": "RecordingData",
+                    "type": "object",
+                    "properties": [
+                        {
+                            "name": "id",
+                            "$ref": "RecordingId"
+                        },
+                        {
+                            "name": "url",
+                            "type": "string"
+                        },
+                        {
+                            "name": "title",
+                            "type": "string"
+                        },
+                        {
+                            "name": "duration",
+                            "type": "integer"
+                        },
+                        {
+                            "name": "lastScreenData",
+                            "type": "string"
+                        },
+                        {
+                            "name": "lastScreenMimeType",
                             "type": "string"
                         }
                     ]
