+++ pdl/protocol.pdl
@@ -8,15 +8,6 @@ domain Recording
   # Globally unique identifier for a recording.
   type RecordingId extends string
 
-  type RecordingData extends object
-    properties
-      RecordingId id
-      string url
-      string title
-      integer duration
-      string lastScreenData 
-      string lastScreenMimeType  
-
   # Unique identifier for the software used to produce a recording.
   type BuildId extends string
 
@@ -1542,7 +1533,7 @@ domain Internal
 
   command setRecordingMetadata
     parameters
-      Recording.RecordingData recordingData
+      RecordingData recordingData
       string authId 
 
   # Add data to a recording. The next message sent after this must be a binary
@@ -1635,3 +1626,14 @@ domain Internal
       string authId
     returns
       array of object recordings
+
+  type RecordingData extends object
+    properties
+      RecordingId id
+      string url
+      string title
+      integer duration
+      string lastScreenData 
+      string lastScreenMimeType  
+
+
