+++ ts/protocol/Internal.ts
@@ -1,7 +1,6 @@
 import {
   BuildId,
   RecordingId,
-  RecordingData,
   TimeStamp
 } from "./Recording";
 import {
@@ -26,6 +25,21 @@ export interface Resource {
    * Checksum of the resource's contents.
    */
   checksum: string;
+}
+
+export interface RecordingData {
+
+  id: RecordingId;
+
+  url: string;
+
+  title: string;
+
+  duration: number;
+
+  lastScreenData: string;
+
+  lastScreenMimeType: string;
 }
 
 export interface createRecordingParameters {
